Move zip-related code to LocalSource
